feat(stock_em_pg): add stock_em_pg interface
add stock_em_pg interface
