+++ docs/changelog.md
@@ -2108,4 +2108,6 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 0.9.28: add: add stock_legu_average_position interface
 
 0.9.29: add: add stock_em_qbzf interface
+
+0.9.30: add: add stock_em_pg interface
 ```
