+++ docs/data/stock/stock.md
@@ -8374,6 +8374,65 @@ print(stock_em_qbzf_df)
 4337  002205   国统股份    -  定向增发  ...   6.92  2011-01-04  2011-01-11  1-3年
 ```
 
+#### 配股
+
+接口: stock_em_pg
+
+目标地址: http://data.eastmoney.com/xg/pg/
+
+描述: 获取东方财富网-数据中心-新股数据-配股
+
+限量: 单次返回所有历史数据
+
+输入参数
+
+| 名称   | 类型 | 必选 | 描述   |
+| -------- | ---- | ---- | --- |
+| - | - | -    | - |
+
+输出参数
+
+| 名称          | 类型 | 默认显示 | 描述           |
+| ------------ | ----- | -------- | ---------------- |
+| 股票代码          | object   | Y        |  -    |
+| 股票简称          | object   | Y        |  -    |
+| 配售代码          | object   | Y        |  -    |
+| 配股数量          | object   | Y        |  注意单位: 股    |
+| 配股比例          | object   | Y        |  -   |
+| 配股价          | object   | Y     |  -    |
+| 最新价          | object   | Y     |  -    |
+| 配股前总股本          | object   | Y     |  注意单位: 股    |
+| 配股后总股本          | object   | Y     |  注意单位: 股    |
+| 股权登记日          | object   | Y     |  -    |
+| 缴款起始日期          | object   | Y     |  -    |
+| 缴款截止日期          | object   | Y     |  -    |
+| 上市日          | object   | Y     |  -    |
+
+接口示例
+
+```python
+import akshare as ak
+stock_em_pg_df = ak.stock_em_pg()
+print(stock_em_pg_df)
+```
+
+数据示例
+
+```
+        股票代码   股票简称    配售代码  ...      缴款起始日期      缴款截止日期         上市日
+0     600888   新疆众和  700888  ...  2021-04-13  2021-04-19  2021-05-07
+1     600919   江苏银行  700919  ...  2020-12-09  2020-12-15  2021-01-14
+2     000728   国元证券  080728  ...  2020-10-14  2020-10-20  2020-10-30
+3     600008   首创股份  700008  ...  2020-09-21  2020-09-25  2020-10-16
+4     002788   鹭燕医药  082788  ...  2020-09-09  2020-09-15  2020-10-13
+      ...    ...     ...  ...         ...         ...         ...
+1031  600651   飞乐音响          ...                          1994-04-25
+1032  600653   申华控股          ...                          1993-06-10
+1033  600652  *ST游久          ...                          1993-03-15
+1034  600609   金杯汽车          ...                          1993-01-29
+1035  600601   ST方科          ...                          1992-06-10
+```
+
 ### 大宗交易
 
 #### 市场统计
